+++ src/main/kotlin/com/dphong/likescash/api/product/controller/ProductController.kt
@@ -1,6 +1,7 @@
-package com.dphong.likescash.api.product
+package com.dphong.likescash.api.product.controller
 
 import com.dphong.likescash.api.product.model.ProductDetails
+import com.dphong.likescash.api.product.service.ProductGetter
 import com.dphong.likescash.common.reponse.DataResult
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
