+++ src/main/kotlin/com/dphong/likescash/api/product/service/ProductGetter.kt
@@ -1,4 +1,4 @@
-package com.dphong.likescash.api.product
+package com.dphong.likescash.api.product.service
 
 import com.dphong.likescash.api.product.model.ProductDetails
 import com.dphong.likescash.common.exception.ResourceNotFoundException
