+++ src/test/kotlin/com/dphong/likescash/mock/TestContainer.kt
@@ -1,6 +1,6 @@
 package com.dphong.likescash.mock
 
-import com.dphong.likescash.api.product.RegisterProductService
+import com.dphong.likescash.api.product.service.RegisterProductService
 import com.dphong.likescash.repository.member.MemberRepository
 import com.dphong.likescash.repository.product.ProductRepository
